Retriever provided at agent creation
