+++ services/chatbot/src/mcpserver/server.py
@@ -5,10 +5,6 @@ import os
 from .config import Config
 import logging
 import time
-from .tool_helpers import (
-    get_any_api_key,
-    build_retrieverQA,
-)
 
 # Configure logging
 logging.basicConfig(
@@ -77,27 +73,6 @@ with open(Config.OPENAPI_SPEC, "r") as f:
 mcp = FastMCP.from_openapi(
     openapi_spec=openapi_spec, client=get_http_client(), name="My crAPI MCP Server"
 )
-
-@mcp.tool(tags={"history", "search", "summary", "context"},)
-async def retriever_tool(question: str) -> str:
-    """Answer questions based on user chat history (summarized and semantically indexed). 
-    Use this when the user asks about prior chats, what they asked earlier, or wants a summary of past conversations.    
-    Answer questions based on the user's prior chat history.
-
-    Use this tool when the user refers to anything mentioned before, asks for a summary of previous messages or sessions, 
-    or references phrases like 'what I said earlier', 'things we discussed', 'my earlier question', 'until now', 'till date', 'all my conversations' or 'previously mentioned'.
-    The chat history is semantically indexed and summarized using vector search."""
-
-    logger.info(f"search_chat_history called with: {question}")
-    api_key=await get_any_api_key()
-    if not api_key:
-        logger.error("API key is not available. Cannot search chat history.")
-        return "OpenAI API key is not available. Cannot search chat history."
-    retrieverQA = build_retrieverQA(api_key=api_key)
-    response = await retrieverQA.ainvoke({"query": question})
-    result = response["result"]
-    logger.info(f"RESULT: {result}")
-    return result
 
 if __name__ == "__main__":
     mcp_server_port = int(os.environ.get("MCP_SERVER_PORT", 5500))
